requirements.txt
google-generativeai
sentence-transformers
PyMuPDF
pandas
python-docx
streamlit
python-dotenv
